Moving audio deprecated interface
