+++ cocos2d/deprecated.js
@@ -667,176 +667,4 @@ if (CC_DEV) {
         },
     });
 
-    (function (audioEngine) {
-        if (typeof audioEngine === 'undefined') return;
-
-        (function (prop) {
-            function error () {
-                cc.error('Sorry, %s.%s is removed.', 'cc.audioEngine', prop);
-            }
-            js.getset(audioEngine, prop, error, error);
-        })('willPlayMusic');
-
-        var musicId = -1;
-        var musicVolume = 1;
-        var effectsVolume = 1;
-        var pauseIDCache = [];
-        js.get(audioEngine, 'playMusic', function () {
-            cc.warn(INFO, 'audioEngine.playMusic', 'audioEngine.play');
-            return function (url, loop) {
-                musicId = audioEngine.play(url, loop, musicVolume);
-                return musicId;
-            }
-        });
-        js.get(audioEngine, 'stopMusic', function () {
-            cc.warn(INFO, 'audioEngine.stopMusic', 'audioEngine.stop');
-            return function () {
-                audioEngine.stop(musicId);
-                return musicId;
-            }
-        });
-        js.get(audioEngine, 'pauseMusic', function () {
-            cc.warn(INFO, 'audioEngine.pauseMusic', 'audioEngine.pause');
-            return function () {
-                audioEngine.pause(musicId);
-                return musicId;
-            }
-        });
-        js.get(audioEngine, 'resumeMusic', function () {
-            cc.warn(INFO, 'audioEngine.resumeMusic', 'audioEngine.resume');
-            return function () {
-                audioEngine.resume(musicId);
-                return musicId;
-            }
-        });
-        js.get(audioEngine, 'rewindMusic', function () {
-            cc.warn(INFO, 'audioEngine.rewindMusic', 'audioEngine.setCurrentTime');
-            return function () {
-                audioEngine.setCurrentTime(musicId, 0);
-                return musicId;
-            }
-        });
-        js.get(audioEngine, 'getMusicVolume', function () {
-            cc.warn(INFO, 'audioEngine.getMusicVolume', 'audioEngine.getVolume');
-            return function () {
-                return musicVolume;
-            }
-        });
-        js.get(audioEngine, 'setMusicVolume', function () {
-            cc.warn(INFO, 'audioEngine.setMusicVolume', 'audioEngine.setVolume');
-            return function (volume) {
-                musicVolume = volume;
-                audioEngine.setVolume(musicId, musicVolume);
-                return musicVolume;
-            }
-        });
-        js.get(audioEngine, 'isMusicPlaying', function () {
-            cc.warn(INFO, 'audioEngine.isMusicPlaying', 'audioEngine.getState');
-            return function () {
-                return audioEngine.getState(musicId) === audioEngine.AudioState.PLAYING;
-            }
-        });
-        js.get(audioEngine, 'playEffect', function () {
-            cc.warn(INFO, 'audioEngine.playEffect', 'audioEngine.play');
-
-            return function (url, loop, volume) {
-                return audioEngine.play(url, loop, volume === undefined ? effectsVolume : volume);
-            }
-        });
-        js.get(audioEngine, 'setEffectsVolume', function (volume) {
-            cc.warn(INFO, 'audioEngine.setEffectsVolume', 'audioEngine.setVolume');
-            return function (volume) {
-                effectsVolume = volume;
-                var id2audio = audioEngine._id2audio;
-                for (var id in id2audio) {
-                    if (id === musicId) continue;
-                    audioEngine.setVolume(id, volume);
-                }
-            }
-        });
-        js.get(audioEngine, 'getEffectsVolume', function () {
-            cc.warn(INFO, 'audioEngine.getEffectsVolume', 'audioEngine.getVolume');
-            return function () {
-                return effectsVolume;
-            }
-        });
-        js.get(audioEngine, 'pauseEffect', function () {
-            cc.warn(INFO, 'audioEngine.pauseEffect', 'audioEngine.pause');
-
-            return function (id) {
-                return audioEngine.pause(id);
-            }
-        });
-        js.get(audioEngine, 'pauseAllEffects', function () {
-            cc.warn(INFO, 'audioEngine.pauseAllEffects', 'audioEngine.pauseAll');
-
-            return function () {
-                var id2audio = audioEngine._id2audio;
-                for (var id in id2audio) {
-                    if (id === musicId) continue;
-                    var audio = id2audio[id];
-                    var state = audio.getState();
-                    if (state === audioEngine.AudioState.PLAYING) {
-                        pauseIDCache.push(id);
-                        audio.pause();
-                    }
-                }
-            }
-        });
-        js.get(audioEngine, 'resumeEffect', function () {
-            cc.warn(INFO, 'audioEngine.resumeEffect', 'audioEngine.resume');
-            return function (id) {
-                audioEngine.resume(id);
-            }
-        });
-        js.get(audioEngine, 'resumeAllEffects', function () {
-            cc.warn(INFO, 'audioEngine.resumeEffect', 'audioEngine.resume');
-            return function () {
-                var id2audio = audioEngine._id2audio;
-                while (pauseIDCache.length > 0) {
-                    var id = pauseIDCache.pop();
-                    var audio = id2audio[id];
-                    if (audio && audio.resume)
-                        audio.resume();
-                }
-            }
-        });
-        js.get(audioEngine, 'stopEffect', function () {
-            cc.warn(INFO, 'audioEngine.stopEffect', 'audioEngine.stop');
-            return function (id) {
-                return audioEngine.stop(id);
-            }
-        });
-        js.get(audioEngine, 'stopAllEffects', function () {
-            cc.warn(INFO, 'audioEngine.stopAllEffects', 'audioEngine.stopAll');
-            return function () {
-                var id2audio = audioEngine._id2audio;
-                for (var id in id2audio) {
-                    if (id === musicId) continue;
-                    var audio = id2audio[id];
-                    var state = audio.getState();
-                    if (state === audioEngine.AudioState.PLAYING) {
-                        audio.stop();
-                    }
-                }
-            }
-        });
-        js.get(audioEngine, 'unloadEffect', function () {
-            cc.warn(INFO, 'audioEngine.unloadEffect', 'audioEngine.stop');
-            return function (id) {
-                return audioEngine.stop(id);
-            }
-        });
-
-        if (!CC_JSB) {
-            js.get(audioEngine, 'end', function () {
-                cc.warn(INFO, 'audioEngine.end', 'audioEngine.stopAll');
-                return function () {
-                    return audioEngine.stopAll();
-                }
-            });
-        }
-
-    })(cc.audioEngine);
-
 }
