+++ jsb/jsb-audio.js
@@ -47,6 +47,10 @@ cc.Audio.Type = {
     cc.audioEngine = audioEngine;
     audioEngine.play = audioEngine.play2d;
     audioEngine.setMaxWebAudioSize = function () {};
+    // deprecated
+    var Module = require('../cocos2d/audio/deprecated');
+    Module.removed(audioEngine);
+    Module.deprecated(audioEngine);
 
     proto.State = audioEngine.AudioState;
 
@@ -102,6 +106,7 @@ cc.Audio.Type = {
     };
 
     proto.preload = function () {
+        this._loaded = true;
         this.emit('load');
     };
